requirements.txt
requests~=2.28.1
pandas~=1.5.2
pytz~=2022.6
tabulate~=0.9.0
APScheduler~=3.9.1.post1
python-dotenv~=0.21.0